LayoutTests/imported/w3c/web-platform-tests/html/cross-origin-embedder-policy/cache-storage-reporting-service-worker.https-expected.txt


FAIL COEP support on ServiceWorker. assert_equals: expected 1 but got 0
FAIL COEP-Report-Only support on ServiceWorker. assert_equals: expected 1 but got 0
